creating an entity
